# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Administrative Subdivisions of Ponorogo Regency (Districts and Villages).",
        "title": "Ponorogo Regency API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Erik Rio Setiawan",
            "url": "http://www.swagger.io/support",
            "email": "erikriosetiawan15@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "ponorogo-api.herokuapp.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/regencies": {
            "get": {
                "description": "Get regencies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "regencies"
                ],
                "summary": "Get Regencies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "regency name search by keyword",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.regenciesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/regencies/{id}": {
            "get": {
                "description": "get regencies by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "regencies"
                ],
                "summary": "Get Regency by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Regency ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.regencyResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.regenciesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Regency"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "controller.regencyResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.Regency"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
        "model.Province": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.Regency": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "province": {
                    "$ref": "#/definitions/model.Province"
                }
            }
        },
...
    }
}